+++ derive-builder-test/tests/basic_setters.rs
@@ -1,14 +1,11 @@
-#[macro_use] extern crate custom_derive;
 #[macro_use] extern crate derive_builder;
 
-custom_derive!{
-    #[derive(Debug, PartialEq, Default, Builder, Clone)]
-    struct Lorem {
-        ipsum: String,
-        pub dolor: Option<String>,
-        pub sit: i32,
-        amet: bool,
-    }
+#[derive(Debug, PartialEq, Default, Builder, Clone)]
+struct Lorem {
+    ipsum: String,
+    pub dolor: Option<String>,
+    pub sit: i32,
+    amet: bool,
 }
 
 impl Lorem {