+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- different setter pattern, e.g. `#[setters(immutable)]`
+- private setters, e.g. `#[setters(private)]`
+- additional debug info via env_logger, e.g. `RUST_LOG=derive_builder=trace cargo test`
+
+### Changed
+- migration to macros 1.1, please refer to the new docs
+
+### Fixed
+- missing lifetime support #21
+
 ## [0.2.1] - 2016-09-24
 
 ### Fixed